refactor(process): migrate forwarders component to v2
