+++ comp/process/bundle.go
@@ -18,7 +18,7 @@ import (
 	"github.com/DataDog/datadog-agent/comp/process/connectionscheck/connectionscheckimpl"
 	"github.com/DataDog/datadog-agent/comp/process/containercheck/containercheckimpl"
 	expvarsfx "github.com/DataDog/datadog-agent/comp/process/expvars/fx"
-	"github.com/DataDog/datadog-agent/comp/process/forwarders/forwardersimpl"
+	forwardersfx "github.com/DataDog/datadog-agent/comp/process/forwarders/fx"
 	gpusubscriber "github.com/DataDog/datadog-agent/comp/process/gpusubscriber/fx"
 	hostinfofx "github.com/DataDog/datadog-agent/comp/process/hostinfo/fx"
 	"github.com/DataDog/datadog-agent/comp/process/processcheck/processcheckimpl"
@@ -56,7 +56,7 @@ func Bundle() fxutil.BundleOptions {
 
 		apiserver.Module(),
 		connectionsforwarderfx.Module(),
-		forwardersimpl.Module(),
+		forwardersfx.Module(),
 		logscompression.Module(),
 
 		gpusubscriber.Module(),
