+++ comp/process/forwarders/def/component.go
